requirements.txt
Django==1.8.1
Markdown==2.6.2
SQLAlchemy==1.0.4
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.10.0
django-postgrespool==0.3.0
django-toolbelt==0.0.1
djangorestframework==3.2.1
gunicorn==19.3.0
psycopg2==2.6
requests==2.3.0
sorl-thumbnail==12.3
static3==0.6.1
whitenoise==1.0.6
wsgiref==0.1.2
